refactor: create ProductCard as a component and use one style for all pages list
